change decimal to floatfield

+++ chain/forms.py
@@ -6,7 +6,7 @@ class InitiateTransactionForm(forms.Form):
     sender_address = forms.CharField(widget=forms.TextInput(attrs={'class' : 'form-control', 'placeholder' : 'Sender address'}))
     sender_private_key = forms.CharField(widget=forms.TextInput(attrs={'class' : 'form-control', 'placeholder' : 'Sender private key'}))
     recipient_address = forms.CharField(widget=forms.TextInput(attrs={'class' : 'form-control', 'placeholder' : 'Recipient address'}))
-    amount_to_send = forms.DecimalField(widget=forms.NumberInput(attrs={'class' : 'form-control', 'step': 0.25, 'placeholder' : 'Amount to send: steps of 0.25'}))
+    amount_to_send = forms.FloatField(widget=forms.NumberInput(attrs={'class' : 'form-control', 'step': 0.25, 'placeholder' : 'Amount to send: steps of 0.25'}))
 
 class InitiateTransactionChoiceFieldForm(forms.Form):
     def __init__(self, *args, **kwargs):
@@ -22,13 +22,13 @@ class InitiateTransactionChoiceFieldForm(forms.Form):
         queryset=BlockAccount.objects.all(),
         required=True,
         widget=forms.Select(attrs={'class' : 'form-control'}))
-    amount_to_send = forms.DecimalField(widget=forms.NumberInput(attrs={'class' : 'form-control', 'step': 0.25}))
+    amount_to_send = forms.FloatField(widget=forms.NumberInput(attrs={'class' : 'form-control', 'step': 0.25}))
 
 class AcceptTransactionForm(forms.Form):
     sender_address = forms.CharField(widget=forms.TextInput(attrs={'class' : 'form-control'}))
     recipient_address = forms.CharField(widget=forms.TextInput(attrs={'class' : 'form-control'}))
     signature = forms.CharField(widget=forms.TextInput(attrs={'class' : 'form-control'}))
-    amount_to_receive = forms.DecimalField(widget=forms.NumberInput(attrs={'class' : 'form-control', 'step': 0.25}))
+    amount_to_receive = forms.FloatField(widget=forms.NumberInput(attrs={'class' : 'form-control', 'step': 0.25}))
 
 class NodeRegistrationForm(forms.Form):
     node_urls = forms.CharField(widget=forms.TextInput(attrs={'class' : 'form-control', 'placeholder' : 'Comma-separated list e.g http://127.0.0.1:5000'}))
